Updated tests for new routings
