--- __tests__/src/pages/covers/[cover_id]/[product_id]/options-coming-soon.test.js
@@ -1,22 +1,0 @@
-import { mockFn } from "@/utils/unit-tests/test-mockup-fn";
-import { render, screen } from "@testing-library/react";
-
-jest.mock("@/common/ComingSoon", () => ({
-  ComingSoon: () => <div data-testid="coming-soon"></div>,
-}));
-
-describe("Options test", () => {
-  const OLD_ENV = process.env;
-  beforeEach(() => {
-    mockFn.useRouter();
-    process.env = { ...OLD_ENV, NEXT_PUBLIC_ENABLE_V2: "false" };
-    const Options =
-      require("@/src/pages/covers/[cover_id]/[product_id]/options").default;
-    render(<Options />);
-  });
-
-  test("Should display coming soon", () => {
-    const comingSoon = screen.getByTestId("coming-soon");
-    expect(comingSoon).toBeInTheDocument();
-  });
-});
