+++ __tests__/src/pages/my-policies/[cover_id]/products/[product_id]/incidents/[timestamp]/claim.test.jsx
@@ -1,0 +1,26 @@
+import { initiateTest } from "@/utils/unit-tests/test-mockup-fn";
+import ClaimPolicyDiversifiedProduct from "@/src/pages/my-policies/[cover_id]/products/[product_id]/incidents/[timestamp]/claim";
+import { screen } from "@testing-library/react";
+
+jest.mock("@/modules/my-policies/ClaimDetailsPage", () => {
+  return {
+    ClaimDetailsPage: ({ disabled }) => {
+      return (
+        <div data-testid="claim-details-page">{disabled && "Coming soon!"}</div>
+      );
+    },
+  };
+});
+
+describe("ClaimPolicyDiversifiedProduct test", () => {
+  const { initialRender } = initiateTest(ClaimPolicyDiversifiedProduct);
+
+  beforeEach(() => {
+    initialRender();
+  });
+
+  test("should display ClaimPolicyDiversifiedProduct component", () => {
+    const policies = screen.getByTestId("claim-details-page");
+    expect(policies).toBeInTheDocument();
+  });
+});
