+++ __tests__/src/pages/reports/[cover_id]/index.test.jsx
@@ -1,0 +1,25 @@
+import { initiateTest } from "@/utils/unit-tests/test-mockup-fn";
+import { screen } from "@testing-library/react";
+import ReportListingPage from "@/src/pages/reports/[cover_id]/index";
+
+jest.mock("@/src/modules/reporting/ReportListing", () => {
+  return {
+    __esModule: true,
+    default: () => {
+      return <div data-testid="listing-table"></div>;
+    },
+  };
+});
+
+describe("ReportListingPage test", () => {
+  const { initialRender } = initiateTest(ReportListingPage);
+
+  beforeEach(() => {
+    initialRender();
+  });
+
+  test("should display ReportListingPage with ReportListing component", () => {
+    const incident = screen.getByTestId("listing-table");
+    expect(incident).toBeInTheDocument();
+  });
+});
