+++ __tests__/src/pages/reports/[cover_id]/products/[product_id]/incidents/[timestamp]/dispute-coming-soon.test.jsx
@@ -1,0 +1,29 @@
+import { initiateTest, mockFn } from "@/utils/unit-tests/test-mockup-fn";
+import { screen } from "@testing-library/react";
+
+jest.mock("@/common/ComingSoon", () => ({
+  ComingSoon: () => <div data-testid="coming-soon"></div>,
+}));
+
+describe("DisputeFormPage test", () => {
+  const OLD_ENV = process.env;
+  beforeEach(() => {
+    process.env = { ...OLD_ENV, NEXT_PUBLIC_FEATURES: "none" };
+    const DisputeFormPage =
+      require("@/src/pages/reports/[cover_id]/products/[product_id]/incidents/[timestamp]/dispute").default;
+
+    const { initialRender } = initiateTest(DisputeFormPage, {}, () => {
+      mockFn.useCoverOrProductData();
+      mockFn.useFetchReport(() => ({
+        data: { incidentReport: false },
+        loading: true,
+      }));
+    });
+    initialRender();
+  });
+
+  test("Should display coming soon", () => {
+    const comingSoon = screen.getByTestId("coming-soon");
+    expect(comingSoon).toBeInTheDocument();
+  });
+});
